Update the pinout map and add 1-Wire to all_together.ino
As we do not need two digital pins for two water temperature
sensors anymore based on 1-Wire protocol, pinout indexes for
digital pins were shifted by one.

The next step is to implement other sensors and create functions
for each of them. OOP might be used as well, but later.